# fix(gc): freeze gc objects after startup 