@burkfers 2025-01-27
I have been trying this branch for about 16 hours of print time and not suffered any TTC (the standard causes such as canbus errors, high neopixel frame rate, throttling host etc have been ruled out).

Will update as time passes.