@ammmze 2025-01-27
Another idea that was given via Discord chat was possibly doing a `gc.freeze()` at the start of a print and `gc.unfreeze()` at the end of the print. My understanding from that conversation is that it would move everything over to the permanent generation, which would then mean gen 2 collection doesn't have to go through as much stuff, which should make it faster. One issue with that is that feature is only available in python 3.7+, so we'd need to do some feature detection to ensure it doesn't completely break for older python versions and people would need to be on 3.7+ if they want this. 

Also I was thinking since there really shouldn't be much that gets loaded once klippy has fully loaded, perhaps we could have klippy do a full gc once everything has finished loading, then just do the `gc.freeze()` and never actually un-freeze them. So then collection would only run on stuff created and dropped since klippy started.