@KevinOConnor 2025-01-28
Interesting, thanks.

I have some high-level thoughts.  These are just some initial thoughts, so don't take them too seriously.  Listed in no particular order.
* We can certainly optimize the Python garbage collection in Klipper.
* The current reporting is in place because we've ran into issues like this in the past.  The code does do a full garbage collection sweep on each restart (see klippy.py).  This could be extended to also run after connect I guess.
* If we limit running of gc based on previous scan times, we risk never running a gc.  It's common for the API clients to setup a "get_status()" watchlist which runs every 250ms.  So, if a gc sweep ever takes more than 250ms then we'll never gc.  Not gc'ing could introduce subtle failures due to swapping and/or OOM issues.
* A good way to avoid gc'ing in Python is to avoid having circular links between objects when those objects are temporary.  Or, if a circular link is needed, then manually break the link prior to dismissing the last reference to it.  For example, see: `git log 3efb424d..1af4a4ae9f`
* I agree it is difficult to get many other developers to think about these types of gc issues.  However, the code will *never* be stable with 3rd party "extras" modifications.  The only way to allow "plugins" in a way that will scale and be stable is to define an API and run those "plugins" in a separate python process.  https://klipper.discourse.group/t/new-proposal-for-klipper-extension-support/19150 
* FYI, even if `self._next_timer` is several hundred milliseconds in the future, it's still not safe for the Python process to become busy for an extended period.  At any point, the mcu could send a message to the host, and the host may need to respond within about 100ms.  (The heater code is 300ms, but other areas of the code could be less.)  New mcu message wakeup the reactor outside of the timer dispatch mechanism.

Again, just some high-level initial thoughts.
-Kevin