@ammmze 2025-01-28
Thanks for sharing your thoughts!

> If we limit running of gc based on previous scan times, we risk never running a gc. It's common for the API clients to setup a "get_status()" watchlist which runs every 250ms. So, if a gc sweep ever takes more than 250ms then we'll never gc. Not gc'ing could introduce subtle failures due to swapping and/or OOM issues.

This is one of my concerns as well. I did try to accommodate it by essentially limiting the tracked gc time to 90% of the highest time between events. Though that 90% selection was arbitrary. But like you pointed out, I can confirm that we never see 300ms between events. I think the max I ever saw was like ~150ms, which means my current code would essentially track the 300ms gen 2 collection as 135ms, meaning it would allow gen 2 collection to run if it sees a 135ms window. And if it ends running that 300ms collection, there is a chance it could still trigger a TTC. So like I mentioned, this solution is not bullet proof...just reduces the chances of a TTC due to gc further.

> The only way to allow "plugins" in a way that will scale and be stable is to define an API and run those "plugins" in a separate python process. https://klipper.discourse.group/t/new-proposal-for-klipper-extension-support/19150

I am intrigued by this and will look forward to more on this. I hadn't see it before. Thanks for sharing!

> FYI, even if self._next_timer is several hundred milliseconds in the future, it's still not safe for the Python process to become busy for an extended period. At any point, the mcu could send a message to the host, and the host may need to respond within about 100ms. (The heater code is 300ms, but other areas of the code could be less.) New mcu message wakeup the reactor outside of the timer dispatch mechanism.

This is a very good point! And i think is a big reason why my current code would only be considered a bandaid. Ultimately, because garbage collection is a blocking process, the only true resolution is to be able to reduce the amount of time it takes to do garbage collection. And I think the separate plugin process will be extremely valuable for that.

In the meantime, I'd still like to try some `gc.collect()` and `gc.freeze()` stuff after klippy `connect` (or freeze on print start and unfreeze on print end/cancel) to see if that will reduce gen 2 collection time. If that proves to be successful, would you be open to accepting code that does some feature detection (i.e. detect if `freeze` is a method on `gc`) and perform that functionality? This would mean this gc optimization would only be available for users using python 3.7+. Users on older python would just skip that functionality.