@ammmze 2025-01-29
Okay...this is seems promising...

I just did a test where i created a new module (`garbage_collection`) that just does the following upon receiving `klippy:connect`:

```python
def handle_connect(self):
    logging.info("Running full garbage collection and freezing")
    gc.collect(0)
    gc.collect(1)
    gc.collect(2)
    gc.freeze()
```

In the reactor, I just have my logging active so i can see the timings before and after. I've only included one set of logs for each gc generation. These are both after restarting klipper and the printer just sitting idle for a few minutes. We can see gen 2 collection is significantly faster if we freeze all the objects present after the initial load of klipper. Now in theory memory from this point _should_ be pretty stable as collection will still run on new stuff getting created. Thanks to @nefelim4ag for pointing me to the `gc.freeze()` functionality.

### BEFORE (i.e. **without** `gc.freeze()`)
```
Running gc gen 0 with 1.05ms freetime. Counts (701, 9, 9)
Finished gc of 716 items in 3.74ms
Running gc gen 1 with 40.69ms freetime. Counts (702, 10, 9)
Finished gc of 814 items in 2.13ms
Running gc gen 2 with 24.96ms freetime. Counts (701, 10, 10)
Finished gc of 1487 items in 326.70ms
```

### AFTER (i.e. **with** `gc.freeze()`)
```
Running gc gen 0 with 40.61ms freetime. Counts (705, 9, 9)
Finished gc of 676 items in 1.69ms
Running gc gen 1 with 40.64ms freetime. Counts (708, 10, 9)
Finished gc of 822 items in 2.45ms
Running gc gen 2 with 39.44ms freetime. Counts (704, 10, 10)
Finished gc of 996 items in 7.65ms
```

Thoughts?

Perhaps this could be an opt-in module initially (i.e. users would need to add `[garbage_collection]` to their config and then eventually if we find it works well for people who are having TTC errors, perhaps we make it enabled by default?